Move struct StatusMessage to 'types' package
In order to avoid cycle depedency.
